requirements.txt
aniso8601==8.0.0
click==7.1.2
Flask==1.1.2
Flask-JWT-Extended==3.24.1
Flask-RESTful==0.3.8
gcloud==0.17.0
googleapis-common-protos==1.51.0
httplib2==0.17.3
itsdangerous==1.1.0
Jinja2==2.11.2
jws==0.1.3
MarkupSafe==1.1.1
oauth2client==3.0.0
protobuf==3.12.0rc1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycryptodome==3.4.3
PyJWT==1.7.1
Pyrebase==3.0.27
python-decouple==3.3
python-jwt==2.0.1
pytz==2020.1
requests==2.11.1
requests-toolbelt==0.7.0
rsa==4.0
six==1.14.0
Werkzeug==1.0.1
